+++ CHANGELOG.md
@@ -135,6 +135,7 @@ All notable changes to this project will be documented in this file.
 - Code generation for HighLowItem (#634)
 - Axis.MinimumRange did not work correctly (#711)
 - FillColor in ErrorColumnSeries (#736)
+- XAxisKey and YAxisKey added to Wpf.Annotations (#743)
 
 ## [2014.1.546] - 2014-10-22
 ### Added